requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
openai
langchain-groq
faiss-cpu
langchain-huggingface